Put file operations in operations.ts file
